helper functions separated,pm add --repo changed to --url
