+++ lib/action.js
@@ -3,146 +3,33 @@ const prompts = require('prompts');
 
 // internal modules
 const path = require('path');
-const fs = require('fs');
 const util = require('util');
 const exec = util.promisify(require('child_process').exec);
-const os = require('os');
-
 
 // helper functions
 const color = require('./colors.js');
-// require('./helpers.js');
 
-// Default settings.
-const settingsData = {
-    "commandToOpen": "code",
-    "projects": []
-}
+const {
+    settings,
+    settingsPath,
+    throwCreateIssueError,
+    writeSettings,
+    openURL,
+    isURL,
+    onCancel,
+    getChoices,
+    selectProject
+} = require('./helper.js');
 
-// Create settings.json if does not exist or just require it if it does exist
-const settingsPath = path.join(os.homedir(), '.projectman', 'settings.json');
+// Includes end.
 
-let settings;
-try {
-    settings = require(settingsPath);
-} catch (err) {
-    const settingsDir = path.join(os.homedir(), '.projectman');
-    if (err.code === 'MODULE_NOT_FOUND') {
-        if (!fs.existsSync(settingsDir)) {
-            fs.mkdirSync(settingsDir);
-        }
-        fs.writeFileSync(settingsPath, JSON.stringify(settingsData, null, 4), 'utf8');
-        settings = settingsData;
-    }
-}
-//
+/*
 
-// Helper functions [START]
-console.error = function (message) {
-    console.log(color.boldRed(">>> ") + message);
-};
+COMMAND: pm open [projectName]
+ARGUMENTS: [projectName] :: Name of the project to open
 
-console.warn = function (message) {
-    console.log(color.boldYellow(">>> ") + message);
-}
+*/
 
-function throwCreateIssueError(err) {
-    console.error("If you think it is my fault please create issue at https://github.com/saurabhdaware/projectman with below log");
-    console.log(color.boldRed("Err:"));
-    console.log(err);
-}
-
-// Takes data as input and writes that data to settings.json
-function writeSettings(data, command = '<command>', successMessage = "Settings updated :D !") {
-    fs.writeFile(settingsPath, JSON.stringify(data, null, 4), err => {
-        if (err) {
-            if (err.code === 'EACCES') {
-                const errCmd = (process.platform == 'win32') ? `an admin` : `a super user ${color.boldYellow(`sudo pm ${command}`)}`;
-                console.error(`Access Denied! please try again as ${errCmd}`);
-                return;
-            }
-            throwCreateIssueError(err);
-            return;
-        }
-        console.log(color.boldGreen(">>> ") + successMessage + color.green(" ✔"));
-    })
-}
-
-async function openURL(url) {
-    let stderr;
-    switch (process.platform) {
-        case "darwin":
-            ({stderr} =  await exec(`open ${url}`));
-            break;
-        case "win32":
-            ({stderr} =  await exec(`start ${url}`));
-            break;
-        default:
-            ({stderr} =  await exec(`xdg-open ${url}`));
-            break;
-    }
-    console.log('here ',stderr)
-    return stderr;
-}
-
-const suggestFilter = (input,choices) => {
-    return Promise.resolve(
-        choices.filter(choice => 
-            choice.title.toLowerCase().includes(input.toLowerCase())
-        )
-    )
-}
-
-
-const onCancel = () => {
-    console.error("See ya ('__') /");
-    process.exit();
-    return false;
-}
-
-function getChoices() {
-    const projects = [...settings.projects];
-    const result = projects.map(({...project}) => { // Spreading project to make it immutable
-        project.title = project.name;
-        project.value = {name: project.name, path: project.path};
-        if (project.editor && project.editor !== settings.commandToOpen) {
-            project.title += color.grey(` (${color.boldGrey('editor:')} ${color.grey(project.editor+')')}`);
-            project.value.editor = project.editor;
-        }
-        return project;
-    });
-
-    return result;
-}
-
-async function selectProject(projectName, action) {
-    let selectedProject;
-    if (!projectName) {
-        // Ask which project he wants to open
-        const questions = [
-            {
-                type: 'autocomplete',
-                message: `Select project to ${action} :`,
-                name: 'selectedProject',
-                choices:getChoices(),
-                limit:40,
-                suggest:suggestFilter
-            }
-        ];
-
-        // Redirecting to stderr in order for it to be used with command substitution
-        ({selectedProject} = await prompts(questions,{onCancel}));
-        
-    } else {
-        // If project name is mentioned then open directly
-        selectedProject = settings.projects.find(project => project.name.toLowerCase() == projectName.toLowerCase());
-    }
-    return selectedProject;
-}
-
-// Helper funcitions [END]
-
-// oncmd: projectman open [projectName]
 async function openProject(projectName) {
     if (settings.projects.length == 0) {
         console.error("No projects to open :(");
@@ -158,24 +45,26 @@ async function openProject(projectName) {
         return;
     }
 
-    const commandToOpen = selectedProject.editor || settings.commandToOpen;
-    let stderr
+    const commandToOpen = selectedProject.editor || settings.commandToOpen; // If project specific editor exists, Then open using that command else use global command
+    let stderr;
     try {
-        if (/^https?:\/\//.test(selectedProject.path)) {
-            const {stderr} = await openURL(selectedProject.path)
+
+        if (isURL(selectedProject.path)) {
+            ({stderr} = await openURL(selectedProject.path)); // If it is URL then open in Browser.
         } else {
-            const {stderr} = await exec(`${commandToOpen}  "${selectedProject.path}"`);
+            ({stderr} = await exec(`${commandToOpen}  "${selectedProject.path}"`)); // This line opens projects
         }
+
         if (stderr) {
             console.error("Could not open project for some reason :(");
             throwCreateIssueError(stderr);
             return;
         }
 
-        console.log(`${color.boldGreen(">>>")} Opening ${selectedProject.name} ${color.green("✔")}`);
+        console.log(`${color.boldGreen(">>>")} Opening ${selectedProject.name} ${color.green("✔")}`); // Success yay!
 
     } catch (err) {
-        console.error("Could not open project :(");
+        console.error("Could not open project :("); // Something broke :(
         console.warn(`Are you sure your editor uses command ${color.yellow(commandToOpen)} to open directories from terminal?`);
         console.warn(`If not, use ${color.yellow('pm seteditor')} to set Editor/IDE of your choice`)
         throwCreateIssueError(err);
@@ -184,26 +73,55 @@ async function openProject(projectName) {
 
 }
 
-// projectman add [projectDirectory]
+/*
+
+COMMAND: pm add [projectDirectory] [--url [link]]
+ARGUMENTS: 
+[projectDirectory] :: Directory of the project that you want to add. OR when --url flag is added, URL of the repository.
+
+FLAG:
+[--url [link]] :: Adds URL.
+
+RETURNS: newProject object.
+
+*/
+
 async function addProject(projectDirectory = '.', cmdObj = undefined) {
     let newProject = {};
     let name;
-    if (cmdObj.repo) {
-        if (projectDirectory === '.') {
-            console.error("Must specify a URL for a repository");
-            return
+    let enteredUrl;
+
+    if (cmdObj.url) {
+        if(projectDirectory !== '.'){
+            console.warn("Project's local directory value will be ignore when --url flag is on");
         }
-        newProject.path = projectDirectory;
-        try {
-            name = projectDirectory.match(/https?:\/\/(.*?)\//)[1];
-        } catch {
-            console.error('Must specify a URL for a repository');
-            return
+
+        if(cmdObj.url == true){
+            ({enteredUrl} = await prompts([{
+                type: 'text',
+                message: 'Project URL :',
+                name: 'enteredUrl',
+                initial: 'https://github.com/',
+                validate:(url) => isURL(url)?true:'Not a valid URL'
+            }],{onCancel}));
+            name = enteredUrl.split('/').pop(); // Get last route of URL to set default name
+            newProject.path = enteredUrl;
+
+        }else{
+            if(!isURL(cmdObj.url)){
+                console.error("Not a valid URL");
+                console.warn("A valid URL looks something like " +color.yellow("https://github.com/saurabhdaware/projectman"));
+                return;
+            }
+            name = cmdObj.url.split('/').pop(); // Get last route of URL to set default name
+            newProject.path = cmdObj.url;
         }
+
     } else {
         newProject.path = path.resolve(projectDirectory);
         name = newProject.path.split(path.sep).pop();
     }
+
 
     ({finalName: newProject.name} = await prompts([{
         type: 'text',
@@ -211,10 +129,6 @@ async function addProject(projectDirectory = '.', cmdObj = undefined) {
         name: 'finalName',
         initial: name
     }],{onCancel}));
-
-    if (cmdObj.repo) {
-        newProject.name = newProject.name.concat(` (${name.split('.')[0]})`)
-    }
 
     if (settings.projects.some(project => project.name.toLowerCase() == newProject.name.toLowerCase())) {
         console.error("Project with this name already exists");
@@ -224,6 +138,8 @@ async function addProject(projectDirectory = '.', cmdObj = undefined) {
     settings.projects.push(newProject);
 
     writeSettings(settings, 'add', "Project Added");
+
+    return newProject;
 }
 
 // pm remove [projectName]
@@ -246,9 +162,10 @@ async function setEditor(command, cmdObj = undefined) {
     let selectedProject;
     let selectedIndex = -1; // this will have the index of selectedProject
     let projectName;
+    const setEditorFilter = project => !isURL(project.path);
     if (cmdObj.forProject) { // --for-project exists
         projectName = (cmdObj.forProject === true) ? undefined : cmdObj.forProject; // if only[--for-project], set undefined else if [--for-project [projectName]] set projectName
-        selectedProject = await selectProject(projectName, 'set editor for'); // sending undefined, calls list of projects to select from
+        selectedProject = await selectProject(projectName, 'set editor for', setEditorFilter); // sending undefined, calls list of projects to select from
         // find the index of selectedProject and add editor to that project
         selectedIndex = settings.projects.findIndex(project => project.name.toLowerCase() == selectedProject.name.toLowerCase())
     }
